#4 #5 Adds Test Runner and Code Coverage
Test Runner was added by default to the project, so
I just had to add Code Coverage. I wrote a quick
dry test to make sure the framework is set up
correctly, and to test Code Coverage. We won't be
tracking coverage reports on the repo for now.

In my documentation, I'll go over how to ensure
Code Coverage is on and working.