#18-21 Implements EventChannel<T> This class should be used as the basis for all event channels. We'll be using concrete 'implementations' to allow us to create these in the Asset Menu by declaring versions of this class with specific types, instead of generic `T`.
For instance:
`public class IntEventChannel : EventChannel<int> {}`
would allow us to use
`[CreateAssetMenu(menuName = "Events/Int Event Channel")]`
as a tag to allow us to create new
channels with that data type as the
argument to the invoking methods.

We will define different kinds of
data type channels in a separate
file.