US322 T325 Implements PropertySpaceData and event channels Adds a bunch of eventchannels to deal with property stuff. It seems like the only logic that needs to happen in these is a simple check for what path to go down, dependant on if there is an owner, if the player owns it, or if another player owns it.
It fires events for Purchase,
Upgrade, or paying a fee to another
player. Inside those event payloads,
the space inserts itself, and has
public functions to modify state if
the manager goes through with
whatever transaction. It also sends
events to UI to show information
on Mouse Hover over a space,
and an event to hide that UI on
exit. It might be better to have
that in the SpaceRenderer, which is
the first line of calling on hover
interactions, but it seems more
clear to keep that responsibility
here.

In the future, it might make sense
to set up a script that automates
attaching the EventChannels so we
don't have to do that manually for
all 40 spaces. Unity Editor scripts
can do that. Alternatively, we can
write a direct path so the classes
automatically make the correct
connections, but that would be
easier to break. I'll look into it
when the rest of the spaces are
implemented and we have a better
idea of what event channels we
actually need.