US444 T472 Adds property buying/rent prototype
This is currently separate from the
strategy implementation, using super
temporary methods in GameManager.cs
right now.

The strategies are fine, but they're
using these parallel data layers that
are redundant to what already exists,
and figuring out how to adapt them
will take longer than we have for
right now.

The goal *right now* is to get
something working, and this works
for our purposes.

Semester 2, we can untangle this.